added pop-up for registration page

+++ clientapplication/src/main/java/com/virtusa/clientapplication/domain/User.java
@@ -12,7 +12,7 @@ public class User {
 
 	private String email;
 
-	private int contact;
+	private Long contact;
 
 	private int enabled;
 
@@ -66,11 +66,13 @@ public class User {
 		this.email = email;
 	}
 
-	public int getContact() {
+	
+
+	public Long getContact() {
 		return contact;
 	}
 
-	public void setContact(int contact) {
+	public void setContact(Long contact) {
 		this.contact = contact;
 	}
